Move privacy helpers into domain modules
