+++ src/logging/log-content.test.ts
