+++ src/logging/log-content.ts
